rename tests file
